Move to monorepo setup
