--- cert-manager/Pulumi.dev.yaml
@@ -1,2 +1,0 @@
-config:
-  cert-manager:kind: "true"
