+++ pkg/v1/certmanager/cert_manager_helm_release.go
@@ -1,4 +1,4 @@
-package release
+package certmanager
 
 import (
 	"github.com/pulumi/pulumi-kubernetes/sdk/v4/go/kubernetes"
