+++ pkg/v1/certmanager/cert_manager_helm_release_common_errors.go
@@ -1,4 +1,4 @@
-package release
+package certmanager
 
 import "fmt"
